Share VA pages between enclaves
Share VA pages between enclaves.  Use a RW semaphore and atomic operations
to protect slots, which allows concurrent slot allocation calls (retry on
collision).  The semaphore only needs to be acquired for writing when the
list itself is being modified.  Add a counter to track the number of used
slots in a VA page to optimize empty/full checks.  Free a VA page when all
slots are emptied.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>