Remove isgx_vma struct and associated list
Remove isgx_vma struct and its associated list, and instead use the
existing kernel functions to find VMAs mapped to an enclave.  Keep
track of the number of VMAs mapped to an enclave so that the driver
can detect when an enclave is no longer in use.

Refactor the functionality of isgx_zap_tcs_ptes to perform a single
walk of the enclave pages instead of walking the VMAs (O(n * log(m))
vs O(n * m)), and move the functionality into isgx_invalidate (the
VMA close callback zaps the the entire VMA).

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>