Lock the enclave for the entire EPC eviction flow
The ETRACK->EWB sequence must be atomic to avoid ETRACK failing due
to PREV_TRK_INCMPL.  The alternative approach is to retry ETRACK in
such a case, but that approach adds complexity without any benefit.
For simplicity and performance, extend the lock to the entirety of
the EBLOCK->ETRACK->EWB sequence.

Releasing the enclave lock after ETRACK leads to a race condition
where a different thread performing eviction on the same enclave
can acquire the lock and attempt ETRACK before the original thread
can reacquire the lock for EWB.  If the enclave has at least one
active thread then the second ETRACK will fail with PREV_TRK_INCMPL.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>