Don't acquire free pages spinlock to check nr free
In the EPC swapping kernel thread, check for a swapping condition,
e.g. free < high, without acquiring the free pages spinlock.  Using
a slightly stale value is acceptable as the algorithm to determine
whether or not swapping needs to be performed does not depend on a
high degree of precision.  Omitting the lock acquisition reduces
the amount of interference from the swapping thread during normal
operation.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>