intel_sgx: remove sgx_(un)pin_mm and call down/up_read(&mmap_sem) directly
Remove sgx_pin_mm and sgx_unpin_mm and instead directly call down_read
and up_read on encl->mm->mmap_sem.  Add a check for SGX_ENCL_DEAD in
sgx_isolate_pages as it previously relied on sgx_pin_mm to check for
the enclave being dead.  The other code in sgx_(un)pin_mm was either
redundant and/or unnecessary.

As both consumers of sgx_pin_mm lock the encl after sgx_pin_mm, it is
more accurate to check SGX_ENCL_DEAD after the consumer locks the encl.
Locking the enclave for the purposes of checking SGX_ENCL_DEAD is not
functionally necessary.

Incrementing mm_count when locking encl->mm->mmap_sem is unnecessary
as mm_count is incremented when encl->mmu_notifier is registered and
is not decremented until the notifier is released in sgx_encl_release.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>