intel_sgx: scan encls in a ctx until EPC pages are isolated
Scan all enclaves in a context until at least one EPC page can be
isolated for swapping.  Only move the selected enclave to the end
of the enclave list, i.e. keep an enclave at the head of the list
if none of its pages could be isolated so it is the first enclave
scanned in the next isolation call.

Add a lock-aware enclave release function for use in the isolation
flow.  A kref to an enclave must be held when acquiring its locks
and that reference is release while holding the context's lock if
no pages could be isolated in the enclave.  This creates deadlock
if sgx_encl_release is invoked since it also tries to acquire the
context's lock.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>