intel_sgx: scan encls in a ctx until EPC pages are isolated
Scan all enclaves in a context until at least one EPC page can be
isolated for swapping.  Track the number of encls in a ctx so that
the scan loop can detect when it has (likely) scanned all encls
without requiring complex logic (the list of encls may be modified
while the scan loop is running).  Add a per-ctx swap epoch cnt and
an associated tracker in the ctx's encls to prevent re-scanning an
encl after it has been scanned in the current or later swap epoch.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>