intel_sgx: kill encl if vm_insert_pfn fails in fault
Update the EPC page tracking immediately after sgx_eldu and kill the
encl if vm_insert_pfn fails.  Previously we tried to EREMOVE the EPC
page if vm_insert_pfn returned an error, but EREMOVE fails if there
are active cpus in the enclave, in which case the driver would lose
track of the EPC page.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>